Undo splitting of `personalAccessKey.js` which was not necessary
